pyproject.toml
[project]
name = "assessment-jpcys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "langsmith>=0.7.13",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
